Use six.ensure_binary to convert strs into bytes in _send_message().
Pull request #17 addressed the issue whereby there were occasional
concatenation problems when message header and mesage body were of
different types. However, the original patch, which uses encode(),
was not compatible with python 2 in cases where the body included
characters that fell outside of the strict set of ASCII characters.